requirement.txt
pip install django

pip install mysql-python

pip install ipdb

pip install django-shell-plus

//ckeditor setup
pip install django-ckeditor( Steps- https://pypi.python.org/pypi/django-ckeditor/4.5.1)

//bootstrap setup
sudo apt-get install nodejs
sudo apt-get install npm
sudo npm install -g bower
bower install bootstrap (in static folder )

//django celery
//run celery beat
celery -A ecampaign  beat -S djcelery.schedulers.DatabaseScheduler
//run worker
celery -A ecampaign worker -l info

//
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
decorator==4.0.4
Django==1.8
django-annoying==0.8.4
django-appconf==1.0.1
django-bower==5.0.4
django-celery==3.1.17
django-ckeditor==5.0.2
django-compressor==1.5
django-countries==3.4.1
django-debug-toolbar==1.3.2
django-shell-plus==1.1.6
future==0.15.2
icalendar==3.9.1
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
kombu==3.0.29
MySQL-python==1.2.5
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
python-dateutil==2.4.2
pytz==2015.7
schedule==0.3.2
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.1.18
traitlets==4.0.0
wheel==0.24.0


